split values template and pass config as interface{}
