+++ internal/template/consts.md.gotmpl
@@ -1,4 +1,7 @@
-{{ define "values" }}
+{{ define "consts" }}
+{{ if . }}
+
+## Constants
 
 {{ range . }}
 
@@ -9,3 +12,4 @@
 {{ end }}
 
 {{ end }}
+{{ end}}
