+++ internal/template/types_consts.md.gotmpl
@@ -1,0 +1,15 @@
+{{ define "typesConsts" }}
+{{ if . }}
+
+#### Constants
+
+{{ range . }}
+
+{{ doc .Doc }}
+
+{{ gocode .Decl.Text }}
+
+{{ end }}
+
+{{ end }}
+{{ end}}
