+++ internal/template/vars.md.gotmpl
@@ -1,0 +1,15 @@
+{{ define "vars" }}
+{{ if . }}
+
+## Variables
+
+{{ range . }}
+
+{{ doc .Doc }}
+
+{{ gocode .Decl.Text }}
+
+{{ end }}
+
+{{ end }}
+{{ end }}
