+++ testdata/pkg10_package_consts_vars/go.mod
@@ -1,3 +1,3 @@
-module pkg6
+module pkg10
 
 go 1.15
